requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
oauthlib==3.2.0
pip==20.3.4
pkg-resources==0.0.0
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
setuptools==44.1.1
tweepy==4.10.0
urllib3==1.26.9
yarl==1.7.2
